requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pillow==10.3.0
numpy==1.26.4
ultralytics==8.2.77
torch==2.3.0
torchvision==0.18.0